requirements.txt
addict
future
lmdb
numpy<2.0
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
timm
torch>=1.7
torchvision
tqdm
yapf
pandas
scikit-learn
openpyxl
xlrd